% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kittenize.R
\docType{data}
\name{base_samples}
\alias{base_samples}
\title{A base of sample images}
\format{An object of class \code{unebase} of length 3.}
\usage{
base_samples
}
\description{
A base of sample images
}
\examples{
  base_samples
}
\keyword{datasets}

