% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kittenize.R
\docType{data}
\name{kittens}
\alias{base_samples}
\alias{kittens}
\title{Bases of sample images}
\format{An object of class \code{unebase} of length 3.}
\usage{
kittens

base_samples
}
\description{
Bases of sample images
}
\examples{
  kittens
  base_samples
}
\keyword{datasets}

