% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{closest_rgb}
\alias{closest_rgb}
\title{For each pixel of X, indicates which base image is the closest (in RGB space)}
\usage{
closest_rgb(X, B)
}
\arguments{
\item{X}{a data frame with columns R, G and B}

\item{B}{a data frame with columns R, G and B}
}
\description{
For each pixel of X, indicates which base image is the closest (in RGB space)
}
\examples{
n <- 200*200
X <- data.frame( R = runif(n),  G = runif(n),  B = runif(n) )
Y <- data.frame( R = runif(50), G = runif(50), B = runif(50) )
system.time( closest_rgb( X, Y) )
}

