% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelize.R, R/tipixel.R
\docType{package}
\name{collage}
\alias{collage}
\alias{collage}
\alias{collage-package}
\title{Image collage}
\usage{
collage(img, tiles = samples, size = 10)
}
\arguments{
\item{img}{image to process}

\item{tiles}{tiles to use, e.g. \code{samples}}

\item{size}{number of pixels, each square of \code{size} times \code{size} is replaced by the closest tile}
}
\description{
Image collage

collage
}
\examples{
\dontrun{
  img <- sample_image()
  collage( file = img, size = 10)
}
}
