% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupsynthpat.R
\name{decoupsynth}
\alias{decoupsynth}
\title{decoupsynth}
\usage{
decoupsynth(img, lig = 10, col = 10, redim = NULL, enhanced = FALSE)
}
\arguments{
\item{img}{array de l 'image}

\item{lig}{nombre de ligne}

\item{col}{nombre de colonne}

\item{redim}{dimension finale de l'image}
}
\description{
decoupe et synthétise une image
}

