% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_rgb}
\alias{dist_rgb}
\title{calculates distances between pixels from X and pixels from B}
\usage{
dist_rgb(X, B)
}
\arguments{
\item{X}{a data frame with columns R, G and B}

\item{B}{a data frame with columns R, G and B}
}
\description{
calculates distances between pixels from X and pixels from B
}
\examples{
n <- 200*200
X <- data.frame( R = runif(n),  G = runif(n),  B = runif(n) )
Y <- data.frame( R = runif(50), G = runif(50), B = runif(50) )
system.time( dist_rgb( X, Y) )
}

