% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_base.R
\name{generate_base}
\alias{generate_base}
\title{makes tiles from jpegs in a directory}
\usage{
generate_base(files, size = 25)
}
\arguments{
\item{files}{image files}

\item{size}{size (width and height) of the generated tiles}
}
\description{
importe le contenu du dossier de tuiles et génère un objet base
}
\examples{
\dontrun{
  images <- jpegs( system.file("base", package = "tipixel" ) )
  generate_base(images, size = 25)
}
}

