% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genere_tuiles.R
\encoding{UTF-8}
\name{generate_mono_tiles}
\alias{generate_mono_tiles}
\title{genere_tuiles}
\usage{
generate_mono_tiles(comb, dir, width = 100, height = 100)
}
\arguments{
\item{comb}{a matrix of data frame of RGB combinations}

\item{dir}{path where to write the tiles}

\item{width}{width of the generated tiles}

\item{height}{height of the generated tiles}
}
\description{
genère des images mono couleurs
}
\examples{
\dontrun{
  comb<-expand.grid(a1=seq(0,1,0.15),a2=seq(0,1,0.15),a3=seq(0,1,0.15))
  dir <- tempfile()
  generate_mono_tiles(lescomb,dir)
}
}

