% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genere_tuiles.R
\encoding{UTF-8}
\name{genere_tuiles}
\alias{genere_tuiles}
\title{genere_tuiles}
\usage{
genere_tuiles(combinaisons, dossier, dim = c(100, 100))
}
\arguments{
\item{combinaisons}{data.frame avec 1 ligne = une combinaison R V B}

\item{dossier}{chemin du dossier de création des tuiles}

\item{dim, }{vecteur de taille 2, dimension en pixel des tuiles}
}
\description{
genère des images mono couleurs
}
\examples{
## Not run
lescomb<-expand.grid(a1=seq(0,1,0.15),a2=seq(0,1,0.15),a3=seq(0,1,0.15))
genere_tuiles(lescomb,dossier='base')
## End(Not run)
}

