% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genere_base.R
\name{genre_base}
\alias{genre_base}
\title{genre_base}
\usage{
genre_base(chemin = "base", redim = NULL)
}
\arguments{
\item{chemin}{le chemin du dosier de tuiles}

\item{redim}{permet de redimensionner les tuiles}
}
\description{
li le contenu du dossier de tuiles et genere un objet base
}
\examples{
les_tuiles <- genre_base("base",redim=c(25,25))

}

