% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelize.R
\name{gridize}
\alias{gridize}
\title{Adds a grid to show the parts of the image to be replaced by tiles}
\usage{
gridize(img, size = 10, grid_col = "white")
}
\arguments{
\item{img}{image}

\item{size}{pixel size}

\item{grid_col}{color to use for the grid}
}
\description{
Adds a grid to show the parts of the image to be replaced by tiles
}
\examples{
\dontrun{
  img <- image_read( sample_image() )
  g <- gridize( img, size = 10, grid_col = "white" )
  plot( as.raster(g) )
}
}

