% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram_luminosity}
\alias{histogram_luminosity}
\alias{histogram_rgb}
\title{luminosity histogram}
\usage{
histogram_luminosity(im)

histogram_rgb(im)
}
\arguments{
\item{im}{an image}
}
\value{
a ggplot2 object
}
\description{
luminosity histogram
}
\examples{
\dontrun{
tigrou <- image_read( "inst/tigrou/tigrou.jpg")
histogram_luminosity(tigrou)
histogram_rgb(tigrou)
}
}
