% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{image_center_crop}
\alias{image_center_crop}
\title{Crops a square in the center of the image}
\usage{
image_center_crop(img)
}
\arguments{
\item{an}{image, typically loaded with \code{\link[magick]{image_read}} or some
other function from \code{magick}}
}
\value{
a square image
}
\description{
Crops a square in the center of the image
}

