% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{image_histogram_data}
\alias{image_histogram_data}
\alias{image_histogram_brightness}
\alias{image_histogram_rgb}
\title{Image histograms}
\usage{
image_histogram_data(im)

image_histogram_brightness(im, coefficients = c(0.3, 0.6, 0.1))

image_histogram_rgb(im)
}
\arguments{
\item{im}{image, such as on read by [image::image_read]}

\item{coefficients}{luma coefficients}
}
\description{
Histograms are a tool used in photography to visualise the tonal
distribution of a picture.
}
\details{
For each channel (red, green, blue), the histogram displays on
the tone (from 0 to 255) on the x axis, and the pixel count on the y axis.

The three channels can be combined together to form the brightness
histogram.

The `image_histogram_data` function makes a `tibble` with 4 columns :
- `tone` : sequence from 0 to 255.
- `red`, `green`, `blue` : pixel count for each channel.

The `image_histogram_brightness` makes a `ggplot` object showing the
brightness histogram.

The `image_histogram_rgb` makes a `ggplot` object showing the histograms
for each channel.
}
\examples{
\dontrun{
tigrou <- image_read( "inst/tigrou/tigrou.jpg")
image_histogram_data(tigrou)
image_histogram_brightness(tigrou)
image_histogram_rgb(tigrou)
}

}
