% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{image_histogram_data}
\alias{image_histogram_data}
\title{Calculates an image histogram}
\usage{
image_histogram_data(im)
}
\arguments{
\item{im}{image}
}
\value{
a data frame of 256 rows with columns `intensity`, `red`, `green`, `blue` and `luminosity`

The `intensity` columns goes from 0 to 255. The `red`, `green` and `blue` column count the number of pixels
in the image with corresponding intensity.

The `luminosity` column is calculated using the luma coefficients (0.3, 0.6 and 0.1) on the
color channels.
}
\description{
Calculates an image histogram
}
