% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_image.R
\name{jpegs}
\alias{jpegs}
\title{get the list of jpegs from a directory}
\usage{
jpegs(path, ...)
}
\arguments{
\item{path}{the directory}

\item{\dots}{passed to \code{\link{list.files}}, mostly used for the \code{recursive} argument}
}
\description{
get the list of jpegs from a directory
}

