% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelize.R
\name{kittenize}
\alias{kittenize}
\title{kittenize}
\usage{
kittenize(...)
}
\arguments{
\item{\dots}{see \code{\link{pixelize}}}
}
\description{
version of \code{\link{pixelize}} using kittens
}
\examples{
\dontrun{
  kittenize(file = sample_image())
}
}

