% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kittenize.R
\docType{data}
\name{kittens}
\alias{kittens}
\title{A base of kittens}
\format{An object of class \code{unebase} of length 3.}
\usage{
kittens
}
\description{
A base of kittens
}
\examples{
  kittens
}
\keyword{datasets}

