% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelize.R
\name{lines_cut}
\alias{lines_cut}
\title{get the right size to split the image into a given number of lines}
\usage{
lines_cut(img, n = 10)
}
\arguments{
\item{img}{an image}

\item{n}{the number of lines}
}
\value{
a number of lines, in addition the \code{lost} attribute
 holds the number of lost pixels at the bottom and the right
}
\description{
get the right size to split the image into a given number of lines
}
