% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mondist.R
\name{mondist}
\alias{mondist}
\title{mondist}
\usage{
mondist(X, B)
}
\arguments{
\item{X}{coordonnée du point}

\item{B}{coordonnée des points de la base}
}
\description{
fonction qui calcule la distance d 'un point par raport a un autre ensemble de point.
}
\examples{
X<-matrix(c(0.5,0.5,0.5))
B<-t(matrix(round(runif(9),2),3))
mondist(X,B)
}

