% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mondist.R
\name{mondist_global}
\alias{mondist_global}
\title{mondist_global}
\usage{
mondist_global(lesX, B, paralell = FALSE, threat = 2)
}
\arguments{
\item{lesX}{les points dont on cherche la distance par rappoirt à la base}

\item{B}{coordonnée des points de la base}

\item{paralell}{booleen si VRAI alors le multiprocessing est utilisé}

\item{threat}{nombre de coeurs à utiliser}
}
\description{
fonction qui calcule la distance d'un ensemble de points par raport à un autre ensemble de points.
}
\examples{
lesX<-do.call(rbind,(rep(list(t(X)),15)))
B<-t(matrix(round(runif(12),2),3))
system.time(mondist_global(lesX,B))
system.time(mondist_global(lesX,B,paralell=TRUE,threat=2))
}

