% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monsize.R
\name{monsize}
\alias{monsize}
\title{monsize}
\usage{
monsize(arr, x, y)
}
\arguments{
\item{chemin}{le chemin du dosier de tuiles}

\item{redim}{permet de redimensionner les tuiles}
}
\description{
redim d'une image en array
}

