% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{pixel}
\alias{pixel}
\title{pixel}
\usage{
pixel(file, lig, col, base, target = NULL)
}
\arguments{
\item{file}{fichier d origine a decouper}

\item{lig}{nombre de ligne}

\item{col}{nombre de colonne}

\item{base}{la base de tuile a utiliser}

\item{target}{fichier de sortie}
}
\description{
fonction a lancer
}
\examples{
## Not run:
lescomb<-expand.grid(a1=seq(0,1,0.15),a2=seq(0,1,0.15),a3=seq(0,1,0.15))
genere_tiles(lescomb,dossier="base")
les_tuiles <- genre_base("base",redim=c(25,25))

A3<-pixel(file="plot.jpg",lig=5,col=5,base=les_tuiles)
A<-pixel(file="plot.jpg",lig=50,col=50,base=bb,target="dessin.jpg")
## End(Not run)
}

