% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelize.R
\name{pixel_grid}
\alias{pixel_grid}
\title{Adds a grid to show the parts of the image to be replaced by tiles}
\usage{
pixel_grid(img, size = 10, grid_col = "white")
}
\arguments{
\item{img}{image}

\item{size}{pixel size}

\item{grid_col}{color to use for the grid}
}
\description{
Adds a grid to show the parts of the image to be replaced by tiles
}
\examples{
\dontrun{
  img <- image_read( sample_image() )

  # a grid of 10x10 tiles
  pixel_grid( img, size = 10, grid_col = "white" )

  # a grid with 10 lines
  pixel_grid( img, size = lines_cut(img, 10) )

  # a grid with 10 columns
  pixel_grid( img, size = columns_cut(img, 10) )
}
}
