% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelize.R
\name{pixel_replace}
\alias{pixel_replace}
\title{pixel_replace}
\usage{
pixel_replace(img, tiles = samples, size = 10)
}
\arguments{
\item{img}{image to process}

\item{tiles}{tiles to use, e.g. \code{samples}}

\item{size}{number of pixels, each square of \code{size} times \code{size} is replaced by the closest tile}
}
\description{
pixel_replace
}
\examples{
\dontrun{
  img <- sample_image()
  pixel_replace( file = img, size = 10)
}
}
