% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelize.R
\name{pixelize}
\alias{pixelize}
\title{pixelize}
\usage{
pixelize(file, base, width = NA, height = NA)
}
\description{
pixelize
}
\examples{
path  <- system.file( "base", package = "tipixel")
tiles <- generate_base(path,size=25)

img <- sample_image()
pixelize( img, tiles, width = 10)
}

