% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelize.R
\name{quality_mask}
\alias{quality_mask}
\title{visual representation of the quality of the tiles}
\usage{
quality_mask(img, tiles = samples, size, min_distance = 0, max_distance)
}
\arguments{
\item{img}{image to pixelize}

\item{tiles}{tile base to use}

\item{size}{size (height and width) of each pixel}

\item{min_distance}{anything below this value is represented in white}

\item{max_distance}{anything above this value is represented in black}
}
\value{
an \code{\link[magick]{imagemagick}} object
}
\description{
splits the image in pixels of \code{size} x \code{size}
}
\details{
for each of these pixels find the best tile in the tile base and retrieve its distance
to the average color of the pixel (distance is between 0 and 1),

The rendered image is a gray scale of these distances.
}
