% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize.R
\encoding{UTF-8}
\name{resize}
\alias{resize}
\title{resize}
\usage{
resize(arr, x, y)
}
\arguments{
\item{arr}{array à redimensionner}

\item{x}{nouvelle dimension en x}

\item{y}{nouvelle dimension en y}
}
\description{
redimensionne une image au format array
}

