% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2048.R
\encoding{UTF-8}
\name{run_2048}
\alias{run_2048}
\title{run_2048}
\usage{
run_2048(nrow, ncol, p = 0.9)
}
\arguments{
\item{nrow}{nomber of row}

\item{ncol}{numver of col}

\item{p}{probability to obtain a 2 (1-p) is the probability to obtain a 4}
}
\description{
The 2048 game
}
\examples{
\dontrun{
run_2048()
}
}

