% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_image.R
\name{sample_image}
\alias{sample_image}
\title{Get the path of a sample image}
\usage{
sample_image(i)
}
\arguments{
\item{i}{index of the sample image, if missing an image is selected at random}
}
\value{
the full path of a sample image
}
\description{
Get the path of a sample image
}

