% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles_animals.R
\name{scrap_animals}
\alias{scrap_animals}
\title{Get urls of animal pictures}
\usage{
scrap_animals(what = "bebe,chats", pages)
}
\arguments{
\item{what}{type of pictures}

\item{pages}{the pages to download}
}
\description{
Get urls of animal pictures
}
