% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinypixel.R
\name{shinycollage}
\alias{shinycollage}
\title{Run shiny app}
\usage{
shinycollage(display.mode = "normal", ...)
}
\arguments{
\item{display.mode}{display mode}

\item{\dots}{see \code{\link[shiny]{runApp}}}
}
\description{
Run shiny app
}
\examples{
\dontrun{
  shinycollage()
}
}
