% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{tiles}
\alias{tiles}
\title{Generates a tile base from a set of files}
\usage{
tiles(files, size = 25L)
}
\arguments{
\item{files}{image files}

\item{size}{size (width and height) of the generated tiles}
}
\value{
a tibble with the columns R, G, B (raw vectors) and tile
}
\description{
Generates a tile base from a set of files
}
\examples{
\dontrun{
  path  <- system.file("base", package = "tipixel" )
  files <- list.files( path, pattern= "jpg$", full.names = TRUE )
  tiles(files, size = 25)
}
}
