% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles_animals.R
\name{tiles_animals}
\alias{tiles_animals}
\title{Tiles from animal pictures}
\usage{
tiles_animals(size = 25, what = "bebe,chats", pages)
}
\arguments{
\item{size}{tile size}

\item{what}{type of pictures}

\item{pages}{the pages to download}
}
\description{
Tiles from animal pictures
}
\examples{
\dontrun{

  kittens <- tiles_animals( size = 50, what = "bebe,chats", pages = 1:20 )
  puppies <- tiles_animals( size = 50, what = "bebe,chiens", pages = 1:20 )
}
}
