% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{tiles_mono}
\alias{tiles_mono}
\title{monochromatic tiles}
\usage{
tiles_mono(colors, size = 25L)
}
\arguments{
\item{colors}{a vector of colors that can be handled by \code{\link[grDevices]{col2rgb}}}

\item{size}{size of generated tiles}
}
\value{
a tiles tibble.
}
\description{
monochromatic tiles
}
\seealso{
\code{\link{tiles}}
}
