% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupsynthpat.R
\name{decoupsynthpath}
\alias{decoupsynthpath}
\title{decoupsynthpath}
\usage{
decoupsynthpath(path, lig = 1, col = 1, redim = NULL)
}
\arguments{
\item{path}{chemin de l'image}

\item{lig}{nombre de ligne}

\item{col}{nombre de colonne}

\item{redim}{dimension finale de l'image}
}
\description{
decoupe et synthétise une image depuis son chemin
}
\examples{
decoupsynthpath("base/base-0.15-0-0.jpg")
}

