% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupsynthpat.R
\encoding{UTF-8}
\name{decoupsynthpath}
\alias{decoupsynthpath}
\title{decoupsynthpath}
\usage{
decoupsynthpath(path, redim = NULL, verbose = FALSE, preload = TRUE)
}
\arguments{
\item{path}{chemin de l'image}

\item{redim}{dimension finale de l'image}

\item{verbose}{booleen rend la fonction bavarde}

\item{preload}{booleen si VRAI les images sont préchargées (prend de la RAM mais accelère le traitement)}
}
\description{
découpe et synthétise une image depuis son chemin
}

