% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_base.R
\name{generate_base}
\alias{generate_base}
\title{Generates a tile base from a set of files}
\usage{
generate_base(files, size = 25L)
}
\arguments{
\item{files}{image files}

\item{size}{size (width and height) of the generated tiles}
}
\description{
importe le contenu du dossier de tuiles et génère un objet base
}
\examples{
\dontrun{
  path   <- system.file("base", package = "tipixel" )
  files <- list.files( path, pattern= "jpg$", full.names = TRUE )
  generate_base(files, size = 25)
}
}
