% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_base.R
\name{generate_base}
\alias{generate_base}
\title{makes tiles from jpegs in a directory}
\usage{
generate_base(path, size = 25, ...)
}
\arguments{
\item{path}{directory where to find the images}

\item{size}{size (width and height) of the generated tiles}

\item{\dots}{used by \code{\link{jpegs}}}
}
\description{
importe le contenu du dossier de tuiles et génère un objet base
}
\examples{
\dontrun{
path <- system.file("base", package = "tipixel" )
generate_base(path, size = 25)
}
}

