% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mono_tiles.R
\name{generate_mono_tiles}
\alias{generate_mono_tiles}
\title{generate monochromatic tiles}
\usage{
generate_mono_tiles(comb, dir, size = 25)
}
\arguments{
\item{comb}{a matrix of data frame of RGB combinations}

\item{dir}{path where to write the tiles}

\item{size}{width and height of the generated tiles}
}
\description{
generate monochromatic tiles
}
\examples{
\dontrun{
  comb<-expand.grid(a1=seq(0,1,0.15),a2=seq(0,1,0.15),a3=seq(0,1,0.15))
  dir <- tempfile()
  generate_mono_tiles(lescomb,dir)
}
}

