% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genere_base.R
\encoding{UTF-8}
\name{genere_base}
\alias{genere_base}
\title{genere_base}
\usage{
genere_base(chemin = "base", redim = NULL, verbose = FALSE,
  preload = TRUE, recursive = TRUE, progress = "text")
}
\arguments{
\item{chemin}{le chemin du dossier de tuiles}

\item{redim}{permet de redimensionner les tuiles}

\item{verbose}{rend la fonction bavarde}

\item{preload}{pre charge les images et rend la base indépendante des fichiers.. mais prend de la place en ram}

\item{recursive}{si VRAI va parcourir les sous dossiers}

\item{progress}{affiche la progression dans la console}
}
\description{
importe le contenu du dossier de tuiles et génère un objet base
}
\examples{
\dontrun{
base <- file.path(find.package('tipixel'),'base')
les_tuiles <- genere_base(base,redim=c(25,25))
genere_base(base)
genere_base(base,redim=c(25,25))
genere_base(base,redim=c(25,25),preload = FALSE)
}
}

