% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridize.R
\name{gridize}
\alias{gridize}
\title{adds a grid to show the parts of the image to be replaced by tiles}
\usage{
gridize(img = readJPEG(file), width = NA, height = NA, size = 1,
  col = "white", average = FALSE, file)
}
\description{
adds a grid to show the parts of the image to be replaced by tiles
}
\examples{
\dontrun{
  img <- sample_image()
  g <- gridize( file = img, width = 10, height = 10, col = "white", size = 2)
  plot( as.raster(g) )
}
}

