% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kittenize.R
\encoding{UTF-8}
\name{kittenize}
\alias{kittenize}
\title{kittenize}
\usage{
kittenize(...)
}
\arguments{
\item{\dots}{see \code{\link{pixelize}}}
}
\description{
Pour ceux qui aiments les chatons
}
\examples{
\dontrun{
library(tipixel)

kittenize(sample_image())
}
}

