% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mondist.R
\encoding{UTF-8}
\name{mondist}
\alias{mondist}
\title{mondist}
\usage{
mondist(X, B)
}
\arguments{
\item{X}{coordonnée du point}

\item{B}{coordonnée des points de la base}
}
\description{
fonction qui calcule la distance d'un point par raport à un autre ensemble de points.
}
\examples{
\dontrun{
X<-matrix(c(0.5,0.5,0.5))
B<-t(matrix(round(runif(9),2),3))
mondist(X,B)
}
}

