% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mondist.R
\name{mondist_global}
\alias{mondist_global}
\title{mondist_global}
\usage{
mondist_global(lesX, B, parallel = FALSE, thread = 2, verbose = TRUE)
}
\arguments{
\item{lesX}{les points dont on cherche la distance par rappoirt à la base}

\item{B}{coordonnée des points de la base}

\item{parallel}{booleen si VRAI alors le multiprocessing est utilisé}

\item{thread}{nombre de coeurs à utiliser}

\item{verbose}{booleen qui rend la fonction bavarde}
}
\description{
fonction qui calcule la distance d'un ensemble de points par raport à un autre ensemble de points.
}
\examples{
\dontrun{
n <- 200*200
lesX <- data.frame( R = runif(n),  G = runif(n),  B = runif(n) )
B    <- cbind( runif(50), runif(50), runif(50) )
system.time(mondist_global(lesX,t(B)))
system.time(mondist_global(lesX,t(B),parallel=TRUE,thread=2))
}
}

