% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\encoding{UTF-8}
\name{pixel}
\alias{pixel}
\title{pixel}
\usage{
pixel(file, lig, col, base, target = NULL, paralell = FALSE, thread = 2,
  open = FALSE, verbose = TRUE, affich = FALSE, doublon = FALSE,
  redim = NULL, random = 1)
}
\arguments{
\item{file}{fichier jpg d'origine à découper}

\item{lig}{nombre de lignes}

\item{col}{nombre de colonnes}

\item{base}{la base de tuile à utiliser}

\item{target}{fichier de sortie en jpg}

\item{paralell}{si oui le calcul en paralell sera utilisé}

\item{thread}{nombre de coeur à utiliser si paralell =TRUE}

\item{open}{si VRAI le fichier target est ouvert en fin de création}

\item{verbose}{si VRAI rend la fonction bavarde}

\item{affich}{booleen si vrai le resultat est affiché en tant que graphique dans R}

\item{doublon}{si FAUX on supprime les doublons avant de calculer les distances}

\item{redim}{utile si la base ne contient pas les images perchargées, précise les dimensions des tuiles}

\item{random}{tire au hasard la vignette parmi les meilleurs tuiles disponibles, random precise ce nombre}
}
\description{
fonction générale du package. Permet de transformer une image en mosaique d'autres images
}
\examples{
\dontrun{
library(tipixel)

base <- file.path(find.package('tipixel'),'base') # un dossier contenant des images d'exemple
img <- sample(list.files(base,full.names = TRUE),1) # on selectionne une image au hasard
plotraster(aperm(jpeg::readJPEG(img),c(2,1,3))) # on affiche cette image
les_tuiles <- genere_base(base,redim=c(25,25)) # on fabrique une base à partir du dossier

pixel(file=img,lig=5,col=5,base=les_tuiles,target='dessin.jpg',open=TRUE,affich = TRUE)
pixel(file=img,lig=50,col=50,base=les_tuiles,target='dessin2.jpg',open=TRUE)
pixel(file=img,lig=100,col=100,base=les_tuiles,target='dessin3.jpg',open=TRUE)
pixel(file=img,lig=200,col=200,base=les_tuiles,target='dessin4.jpg'
,open=TRUE,paralell = TRUE,thread = 2) #on utilise 2 coeur sur la machine pour accelerer le processus

lescomb<-expand.grid(a1=seq(0,1,0.15),a2=seq(0,1,0.15),a3=seq(0,1,0.15))
genere_tuiles(lescomb,dossier='my_pict') # on genere des tuiles d'une seule couleur
les_tuiles2 <- genere_base('my_pict')
pixel(file=img,lig=100,col=100,base=les_tuiles2,target='dessin4.jpg',open=TRUE)

}
}

