% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelize.R
\name{pixelize}
\alias{pixelize}
\title{pixelize}
\usage{
pixelize(img = readJPEG(file), base = base_samples, width = NA,
  height = NA, file)
}
\description{
pixelize
}
\examples{
img <- sample_image()
pixelize( file = img, width = 10)
}

