% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelize.R
\name{pixelize}
\alias{pixelize}
\title{pixelize}
\usage{
pixelize(file, base, width = NA, height = NA)
}
\description{
pixelize
}
\examples{
base <- system.file( "base", package = "tipixel")
les_tuiles <- genere_base(base,redim=c(25,25), progress = "text" )
img <- sample_image()
pixelize( img, les_tuiles, width = 10)
}

