% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelize.R
\name{pixelize}
\alias{pixelize}
\title{pixelize}
\usage{
pixelize(img, base = base_samples, size = 10)
}
\arguments{
\item{img}{image to process}

\item{base}{tile base, e.g. \code{base_samples}}

\item{size}{number of pixels, each square of \code{size} times \code{size} is replaced by the
closest tile from \code{base}}
}
\description{
pixelize
}
\examples{
\dontrun{
  img <- sample_image()
  pixelize( file = img, size = 10)
}
}
