% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelize.R
\name{pixelize}
\alias{pixelize}
\title{pixelize}
\usage{
pixelize(img, base = base_samples, width = NA, height = NA)
}
\description{
pixelize
}
\examples{
\dontrun{
  img <- sample_image()
  pixelize( file = img, width = 10)
}
}

