% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotraster.R
\encoding{UTF-8}
\name{plotraster}
\alias{plotraster}
\title{plotraster}
\usage{
plotraster(img, verbose = TRUE)
}
\arguments{
\item{img}{l 'image (en array)}
}
\description{
affiche une image raster dans la fenêtre graphique
}

