% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize.R
\name{resize}
\alias{resize}
\title{resize}
\usage{
resize(arr, x, y)
}
\arguments{
\item{arr}{array}

\item{x}{nouvelle dimension en x}

\item{y}{nouvelle dimension en y}
}
\description{
redim d'une image en array
}

