% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bases.R
\docType{data}
\name{samples}
\alias{samples}
\alias{samples}
\alias{kittens}
\alias{samples}
\alias{puppies}
\alias{samples}
\alias{useR2017}
\alias{samples}
\alias{emojis}
\title{sample tiles}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 54 rows and 5 columns.}
\usage{
samples

kittens

puppies

useR2017

emojis
}
\description{
sample tiles
}
\keyword{datasets}
